@@ -22,12 +22,13 @@
         "packages/cpt-ui",
         "packages/cdk",
         "packages/cognito",
-        "packages/prescriptionSearchLambda",
+        "packages/prescriptionListLambda",
         "packages/common/middyErrorHandler",
         "packages/common/testing",
         "packages/common/authFunctions",
         "packages/auth_demo",
         "packages/trackerUserInfoLambda",
+        "packages/CIS2SignOutLambda",
         "packages/selectedRoleLambda"
     ],
     "devDependencies": {
@@ -37,29 +38,30 @@
         "@testing-library/jest-dom": "^6.6.2",
         "@testing-library/react": "^16.2.0",
         "@types/jest": "^29.5.14",
-        "@types/node": "^22.13.9",
+        "@types/node": "^22.13.14",
         "@types/react": "18.3.12",
-        "@typescript-eslint/eslint-plugin": "^8.26.0",
-        "@typescript-eslint/parser": "^8.26.0",
+        "@typescript-eslint/eslint-plugin": "^8.27.0",
+        "@typescript-eslint/parser": "^8.28.0",
         "aws-sdk-client-mock": "^4.1.0",
-        "eslint": "^9.21.0",
-        "eslint-config-prettier": "^10.0.2",
+        "eslint": "^9.23.0",
+        "eslint-config-prettier": "^10.1.1",
         "eslint-plugin-import-newlines": "^1.3.4",
-        "eslint-plugin-prettier": "^5.2.3",
+        "eslint-plugin-prettier": "^5.2.5",
+        "globals": "^16.0.0",
         "jest": "^29.7.0",
         "jest-environment-jsdom": "^29.7.0",
         "jest-junit": "^16.0.0",
         "license-checker": "^25.0.1",
         "semantic-release": "^24.2.3",
-        "ts-jest": "^29.2.6",
+        "ts-jest": "^29.3.0",
         "ts-node": "^10.9.2",
         "typescript": "^5.8.2",
-        "typescript-eslint": "^8.26.0"
+        "typescript-eslint": "^8.28.0"
     },
     "dependencies": {
         "conventional-changelog-eslint": "^6.0.0",
-        "esbuild": "^0.25.0",
-        "next": "^15.2.1",
+        "esbuild": "^0.25.1",
+        "next": "^15.2.4",
         "nhsuk-react-components": "^5.0.0",
         "react": "^18.3.1",
         "react-dom": "^18.3.1"
