@@ -6,15 +6,15 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "lint": "eslint . --report-unused-disable-directives --max-warnings 0",
+    "lint": "eslint  --max-warnings 0 --fix --config ../../eslint.config.mjs .",
     "start": "vite preview",
     "test": "NODE_NO_WARNINGS=1 jest --no-cache --coverage",
     "check-licenses": "license-checker --failOn GPL --failOn LGPL --start ../.."
   },
   "dependencies": {
-    "aws-amplify": "^6.13.2",
-    "axios": "^1.8.1",
-    "i18next": "^24.2.1",
+    "aws-amplify": "^6.14.1",
+    "axios": "^1.8.4",
+    "i18next": "^24.2.3",
     "i18next-browser-languagedetector": "^8.0.4",
     "nhsuk-frontend": "^8.0.2",
     "i18next-http-backend": "^3.0.2",
@@ -22,26 +22,26 @@
     "react": "^18",
     "react-dom": "^18",
     "react-i18next": "^15.4.1",
-    "react-intl": "^7.0.1",
-    "sass": "^1.85.1",
+    "react-intl": "^7.1.10",
+    "sass": "^1.86.0",
     "uuid": "^11.1.0",
     "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
-    "@swc/core": "^1.11.6",
+    "@swc/core": "^1.11.13",
     "@swc/jest": "^0.2.37",
     "@testing-library/jest-dom": "^6.6.2",
     "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.14",
-    "@types/node": "^22.13.9",
+    "@types/node": "^22.13.14",
     "@types/react": "^18.2.64",
     "@types/react-dom": "^18.2.21",
-    "@typescript-eslint/eslint-plugin": "^8.26.0",
-    "@typescript-eslint/parser": "^8.26.0",
+    "@typescript-eslint/eslint-plugin": "^8.27.0",
+    "@typescript-eslint/parser": "^8.28.0",
     "@vitejs/plugin-react": "^4.2.1",
-    "@vitejs/plugin-react-swc": "^3.7.2",
-    "eslint": "^9.21.0",
+    "@vitejs/plugin-react-swc": "^3.8.1",
+    "eslint": "^9.23.0",
     "eslint-plugin-jest": "^28.11.0",
     "eslint-plugin-react": "^7.37.4",
     "eslint-plugin-react-hooks": "^5.2.0",
@@ -50,9 +50,9 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "jest-watch-typeahead": "^2.2.2",
-    "sass": "^1.85.1",
-    "ts-jest": "^29.2.6",
+    "sass": "^1.86.0",
+    "ts-jest": "^29.3.0",
     "typescript": "^5.8.2",
-    "vite": "^6.2.0"
+    "vite": "^6.2.3"
   }
 }
