@@ -23,6 +23,8 @@ export interface CloudfrontBehaviorsProps {
   readonly stackName: string
   readonly apiGatewayOrigin: RestApiOrigin
   readonly apiGatewayRequestPolicy: OriginRequestPolicy
+  readonly oauth2GatewayOrigin: RestApiOrigin
+  readonly oauth2GatewayRequestPolicy: OriginRequestPolicy
   readonly staticContentBucketOrigin: IOrigin
 }
 
@@ -61,6 +63,10 @@ export class CloudfrontBehaviors extends Construct{
         {
           key: "api_path",
           value: "/api"
+        },
+        {
+          key: "oauth2_proxyPath",
+          value: "/oauth2"
         },
         {
           key: "jwks_rewrite",
@@ -147,6 +153,21 @@ export class CloudfrontBehaviors extends Construct{
     on how many can be created simultaneously */
     apiGatewayStripPathFunction.node.addDependency(s3StaticContentUriRewriteFunction)
 
+    const oauth2GatewayStripPathFunction = new CloudfrontFunction(this, "OAuth2GatewayStripPathFunction", {
+      functionName: `${props.serviceName}-OAuth2GatewayStripPathFunction`,
+      sourceFileName: "genericStripPathUriRewrite.js",
+      keyValueStore: keyValueStore,
+      codeReplacements: [
+        {
+          valueToReplace: "PATH_PLACEHOLDER",
+          replacementValue: "oauth2_proxyPath"
+        }
+      ]
+    })
+    /* Add dependency on previous function to force them to build one by one to avoid aws limits
+    on how many can be created simultaneously */
+    oauth2GatewayStripPathFunction.node.addDependency(apiGatewayStripPathFunction)
+
     const s3JwksUriRewriteFunction = new CloudfrontFunction(this, "s3JwksUriRewriteFunction", {
       functionName: `${props.serviceName}-s3JwksUriRewriteFunction`,
       sourceFileName: "genericS3FixedObjectUriRewrite.js",
@@ -160,7 +181,7 @@ export class CloudfrontBehaviors extends Construct{
     })
     /* Add dependency on previous function to force them to build one by one to avoid aws limits
     on how many can be created simultaneously */
-    s3JwksUriRewriteFunction.node.addDependency(apiGatewayStripPathFunction)
+    s3JwksUriRewriteFunction.node.addDependency(oauth2GatewayStripPathFunction)
 
     // eslint-disable-next-line max-len
     const authDemoStaticContentUriRewriteFunction = new CloudfrontFunction(this, "authDemoStaticContentUriRewriteFunction", {
@@ -203,6 +224,19 @@ export class CloudfrontBehaviors extends Construct{
         functionAssociations: [
           {
             function: apiGatewayStripPathFunction.function,
+            eventType: FunctionEventType.VIEWER_REQUEST
+          }
+        ]
+      },
+      "/oauth2/*": {
+        origin: props.oauth2GatewayOrigin,
+        allowedMethods: AllowedMethods.ALLOW_ALL,
+        viewerProtocolPolicy: ViewerProtocolPolicy.REDIRECT_TO_HTTPS,
+        originRequestPolicy: props.oauth2GatewayRequestPolicy,
+        cachePolicy: CachePolicy.CACHING_DISABLED,
+        functionAssociations: [
+          {
+            function: oauth2GatewayStripPathFunction.function,
             eventType: FunctionEventType.VIEWER_REQUEST
           }
         ]
