@@ -24,21 +24,23 @@ export interface DynamodbProps {
 }
 
 /**
- * Dynamodb table used for user state information
+ * Dynamodb tables used for user state information
  */
-
 export class Dynamodb extends Construct {
   public readonly tokenMappingTable: TableV2
   public readonly useTokensMappingKmsKeyPolicy: ManagedPolicy
   public readonly tokenMappingTableWritePolicy: ManagedPolicy
   public readonly tokenMappingTableReadPolicy: ManagedPolicy
+  //
+  public readonly stateMappingTable: TableV2
+  public readonly useStateMappingKmsKeyPolicy: ManagedPolicy
+  public readonly stateMappingTableWritePolicy: ManagedPolicy
+  public readonly stateMappingTableReadPolicy: ManagedPolicy
 
   public constructor(scope: Construct, id: string, props: DynamodbProps) {
     super(scope, id)
 
-    // Resources
-
-    // kms key for the table
+    // KMS key for token mapping table
     const tokensMappingKmsKey = new Key(this, "TokensMappingKMSKey", {
       removalPolicy: RemovalPolicy.DESTROY,
       pendingWindow: Duration.days(7),
@@ -80,7 +82,7 @@ export class Dynamodb extends Construct {
       })
     })
 
-    // the table
+    // Token Mapping Table
     const tokenMappingTable = new TableV2(this, "TokenMappingTable", {
       partitionKey: {
         name: "username",
@@ -92,10 +94,9 @@ export class Dynamodb extends Construct {
       encryption: TableEncryptionV2.customerManagedKey(tokensMappingKmsKey),
       billing: Billing.onDemand(),
       timeToLiveAttribute: "ExpiryTime"
-
     })
 
-    // policy to use kms key
+    // Policy to use token mapping KMS key
     const useTokensMappingKmsKeyPolicy = new ManagedPolicy(this, "UseTokensMappingKMSKeyPolicy", {
       statements: [
         new PolicyStatement({
@@ -150,10 +151,126 @@ export class Dynamodb extends Construct {
       ]
     })
 
-    // Outputs
+    // KMS key for state mapping table
+    const stateMappingKmsKey = new Key(this, "StateMappingKMSKey", {
+      removalPolicy: RemovalPolicy.DESTROY,
+      pendingWindow: Duration.days(7),
+      alias: `${props.stackName}-StateMappingKMSKey`,
+      description: `${props.stackName}-StateMappingKMSKey`,
+      enableKeyRotation: true,
+      policy: new PolicyDocument({
+        statements: [
+          new PolicyStatement({
+            sid: "Enable IAM User Permissions",
+            effect: Effect.ALLOW,
+            actions: [
+              "kms:*"
+            ],
+            principals: [
+              new AccountRootPrincipal //NOSONAR: Access is controlled via ARN condition.
+            ],
+            resources: ["*"]
+          }),
+          new PolicyStatement({
+            sid: "Enable read only decrypt",
+            effect: Effect.ALLOW,
+            actions: [
+              "kms:DescribeKey",
+              "kms:Decrypt"
+            ],
+            principals: [
+              new AnyPrincipal()
+            ],
+            resources: ["*"],
+            conditions: {
+              ArnLike: {
+                // eslint-disable-next-line max-len
+                "aws:PrincipalArn": `arn:aws:iam::${props.account}:role/aws-reserved/sso.amazonaws.com/${props.region}/AWSReservedSSO_ReadOnly*`
+              }
+            }
+          })
+        ]
+      })
+    })
+
+    // State Mapping Table
+    const stateMappingTable = new TableV2(this, "StateMappingTable", {
+      partitionKey: {
+        name: "State",
+        type: AttributeType.STRING
+      },
+      tableName: `${props.stackName}-StateMapping`,
+      removalPolicy: RemovalPolicy.DESTROY,
+      pointInTimeRecovery: true,
+      encryption: TableEncryptionV2.customerManagedKey(stateMappingKmsKey),
+      billing: Billing.onDemand(),
+      timeToLiveAttribute: "ExpiryTime"
+    })
+
+    // Policy to use state mapping KMS key
+    const useStateMappingKmsKeyPolicy = new ManagedPolicy(this, "UseStateMappingKMSKeyPolicy", {
+      statements: [
+        new PolicyStatement({
+          actions: [
+            "kms:DescribeKey",
+            "kms:GenerateDataKey",
+            "kms:Encrypt",
+            "kms:ReEncryptFrom",
+            "kms:ReEncryptTo",
+            "kms:Decrypt"
+          ],
+          resources: [
+            stateMappingKmsKey.keyArn
+          ]
+        })
+      ]
+    })
+
+    const stateTableReadManagedPolicy = new ManagedPolicy(this, "StateTableReadManagedPolicy", {
+      statements: [
+        new PolicyStatement({
+          actions: [
+            "dynamodb:GetItem",
+            "dynamodb:BatchGetItem",
+            "dynamodb:Scan",
+            "dynamodb:Query",
+            "dynamodb:ConditionCheckItem",
+            "dynamodb:DescribeTable"
+          ],
+          resources: [
+            stateMappingTable.tableArn,
+            `${stateMappingTable.tableArn}/index/*`
+          ]
+        })
+      ]
+    })
+
+    const stateTableWriteManagedPolicy = new ManagedPolicy(this, "StateTableWriteManagedPolicy", {
+      statements: [
+        new PolicyStatement({
+          actions: [
+            "dynamodb:PutItem",
+            "dynamodb:BatchWriteItem",
+            "dynamodb:UpdateItem",
+            "dynamodb:DeleteItem"
+          ],
+          resources: [
+            stateMappingTable.tableArn,
+            `${stateMappingTable.tableArn}/index/*`
+          ]
+        })
+      ]
+    })
+
+    // Outputs: assign the created resources to the class properties
     this.tokenMappingTable = tokenMappingTable
     this.useTokensMappingKmsKeyPolicy = useTokensMappingKmsKeyPolicy
     this.tokenMappingTableWritePolicy = tableWriteManagedPolicy
     this.tokenMappingTableReadPolicy = tableReadManagedPolicy
+
+    this.stateMappingTable = stateMappingTable
+    this.useStateMappingKmsKeyPolicy = useStateMappingKmsKeyPolicy
+    this.stateMappingTableWritePolicy = stateTableWriteManagedPolicy
+    this.stateMappingTableReadPolicy = stateTableReadManagedPolicy
   }
 }
