@@ -1,12 +1,21 @@
-import React, { useState, useEffect, } from "react"
-import { useNavigate } from "react-router-dom"
-import { Container, Col, Row, Details, Table, ErrorSummary, InsetText } from "nhsuk-react-components"
+import React, {useState, useEffect} from "react"
+import {useNavigate} from "react-router-dom"
+import {
+  Container,
+  Col,
+  Row,
+  Details,
+  Table,
+  ErrorSummary,
+  InsetText
+} from "nhsuk-react-components"
 
-import { useAccess } from "@/context/AccessProvider"
+import {useAccess} from "@/context/AccessProvider"
 import EpsCard from "@/components/EpsCard"
 import EpsSpinner from "@/components/EpsSpinner"
-import { RoleDetails } from "@/types/TrackerUserInfoTypes"
-import { Button } from "./ReactRouterButton"
+import {RoleDetails} from "@/types/TrackerUserInfoTypes"
+import {Button} from "./ReactRouterButton"
+import {FRONTEND_PATHS} from "@/constants/environment"
 
 // This is passed to the EPS card component.
 export type RolesWithAccessProps = {
@@ -50,7 +59,7 @@ interface RoleSelectionPageProps {
 }
 
 export default function RoleSelectionPage({
-  contentText,
+  contentText
 }: RoleSelectionPageProps) {
   const {
     title,
@@ -67,45 +76,60 @@ export default function RoleSelectionPage({
     rolesWithoutAccessHeader,
     noODSCode,
     noRoleName,
-    errorDuringRoleSelection,
+    errorDuringRoleSelection
   } = contentText
 
   const {
     noAccess,
     selectedRole,
-    rolesWithAccess: rawRolesWithAccess = [],
-    rolesWithoutAccess: rawRolesWithoutAccess = [],
+    updateSelectedRole,
+    rolesWithAccess,
+    rolesWithoutAccess,
     loading,
-    error,
+    error
   } = useAccess()
 
   const [loginInfoMessage, setLoginInfoMessage] = useState<string | null>(null)
   const [redirecting, setRedirecting] = useState<boolean>(false)
   const navigate = useNavigate()
 
-  // Transform roles data for display
-  const rolesWithAccess = !noAccess
-    ? rawRolesWithAccess.map((role: RoleDetails, index) => ({
-      uuid: `role_with_access_${index}`,
-      role,
-      link: "/yourselectedrole",
-    }))
-    : [];
+  const [roleCardPropsWithAccess, setRoleCardPropsWithAccess] = useState<Array<RolesWithAccessProps>>([])
+  const [roleCardPropsWithoutAccess, setRoleCardPropsWithoutAccess] = useState<Array<RolesWithoutAccessProps>>([])
 
-  const rolesWithoutAccess = rawRolesWithoutAccess.map((role, index) => ({
-    uuid: `role_without_access_${index}`,
-    roleName: role.role_name || noRoleName,
-    orgName: role.org_name || noOrgName,
-    odsCode: role.org_code || noODSCode,
-  }))
+  useEffect(() => {
+    // Transform roles data for display
+    setRoleCardPropsWithAccess((!noAccess)
+      ? rolesWithAccess.map((role: RoleDetails, index) => ({
+        uuid: `role_with_access_${index}`,
+        role,
+        link: FRONTEND_PATHS.SELECTED_ROLE
+      }))
+      : []
+    )
+
+    setRoleCardPropsWithoutAccess(rolesWithoutAccess.map((role, index) => ({
+      uuid: `role_without_access_${index}`,
+      roleName: role.role_name || noRoleName,
+      orgName: role.org_name || noOrgName,
+      odsCode: role.org_code || noODSCode
+    })))
+
+    console.warn("RoleCardPropsWithAccess length: ", {roleCardPropsWithAccess, loading, error})
+  }, [rolesWithAccess, rolesWithoutAccess])
 
   // Handle auto-redirect for single role
   useEffect(() => {
-    if (rawRolesWithAccess.length === 1 && rawRolesWithoutAccess.length === 0) {
+    if (rolesWithAccess.length === 1 && rolesWithoutAccess.length === 0) {
       setRedirecting(true)
-      navigate("/searchforaprescription")
+      updateSelectedRole(rolesWithAccess[0])
+        .then(() => {
+          navigate(FRONTEND_PATHS.SEARCH)
+        })
+        .catch((err) => {
+          console.error(err)
+        })
     }
-  }, [rawRolesWithAccess, rawRolesWithoutAccess, navigate])
+  }, [rolesWithAccess, rolesWithoutAccess, navigate])
 
   // Set login message when selected role is available
   useEffect(() => {
@@ -113,10 +137,11 @@ export default function RoleSelectionPage({
 
     if (!loginInfoMessage && selectedRole) {
       setLoginInfoMessage(
-        `You are currently logged in at ${selectedRole.org_name || noOrgName} (ODS: ${selectedRole.org_code || noODSCode}) with ${selectedRole.role_name || noRoleName}.`
-      );
+        `You are currently logged in at ${selectedRole.org_name || noOrgName} ` +
+        `(ODS: ${selectedRole.org_code || noODSCode}) with ${selectedRole.role_name || noRoleName}.`
+      )
     }
-  }, [selectedRole, loginInfoMessage, noOrgName, noODSCode, noRoleName, loading]);
+  }, [selectedRole, loginInfoMessage, noOrgName, noODSCode, noRoleName, loading])
 
   // Show spinner while loading or redirecting
   if (loading || redirecting) {
@@ -175,7 +200,7 @@ export default function RoleSelectionPage({
                 </span>
                 <span className="nhsuk-caption-l nhsuk-caption--bottom">
                   <span className="nhsuk-u-visually-hidden"> - </span>
-                  {!noAccess && caption}
+                  {(!noAccess) && caption}
                 </span>
               </span>
             </h1>
@@ -190,7 +215,7 @@ export default function RoleSelectionPage({
                   </span>
                   {loginInfoMessage && (
                     <p
-                      dangerouslySetInnerHTML={{ __html: loginInfoMessage }}
+                      dangerouslySetInnerHTML={{__html: loginInfoMessage}}
                     ></p>
                   )}
                 </InsetText>
@@ -205,11 +230,11 @@ export default function RoleSelectionPage({
             )}
           </Col>
 
-          {!noAccess && rolesWithAccess.length > 0 && (
+          {(!noAccess) && (roleCardPropsWithAccess.length > 0) && (
             <Col width="two-thirds">
               <div className="section">
-                {rolesWithAccess.map((role: RolesWithAccessProps) => (
-                  <EpsCard {...role} key={role.uuid} />
+                {roleCardPropsWithAccess.map((roleCardProps: RolesWithAccessProps) => (
+                  <EpsCard {...roleCardProps} key={roleCardProps.uuid} />
                 ))}
               </div>
             </Col>
@@ -230,7 +255,7 @@ export default function RoleSelectionPage({
                     </Table.Row>
                   </Table.Head>
                   <Table.Body>
-                    {rolesWithoutAccess.map(
+                    {roleCardPropsWithoutAccess.map(
                       (roleItem: RolesWithoutAccessProps) => (
                         <Table.Row key={roleItem.uuid}>
                           <Table.Cell data-testid="change-role-name-cell">
