@@ -1,6 +1,6 @@
-import React from 'react'
-import { Button as NHSButton } from 'nhsuk-react-components'
-import { useNavigate } from 'react-router-dom'
+import React from "react"
+import {Button as NHSButton} from "nhsuk-react-components"
+import {useNavigate} from "react-router-dom"
 
 interface ReactRouterButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
     /** The text content of the button */
@@ -12,41 +12,41 @@ interface ReactRouterButtonProps extends React.ButtonHTMLAttributes<HTMLButtonEl
     /** Optional additional className */
     className?: string
     /** Data test id for testing */
-    'data-testid'?: string
+    "data-testid"?: string
 }
 
-
 export const Button: React.FC<ReactRouterButtonProps> = ({
-    children,
-    to,
-    disabled = false,
-    onClick,
-    className = '',
-    'data-testid': testId,
-    ...props
+  children,
+  to,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  disabled = false,
+  onClick,
+  className = "",
+  "data-testid": testId,
+  ...props
 }) => {
-    const navigate = useNavigate()
+  const navigate = useNavigate()
 
-    const handleClick = (event: React.MouseEvent<HTMLButtonElement> | React.KeyboardEvent<HTMLButtonElement>) => {
-        if (to) {
-            event.preventDefault()
-            const absolutePath = to.startsWith('/') ? to : `/${to}`
-            navigate(absolutePath)
-            return
-        }
-
-        onClick?.(event)
+  const handleClick = (event: React.MouseEvent<HTMLButtonElement> | React.KeyboardEvent<HTMLButtonElement>) => {
+    if (to) {
+      event.preventDefault()
+      const absolutePath = to.startsWith("/") ? to : `/${to}`
+      navigate(absolutePath)
+      return
     }
 
-    return (
-        <NHSButton
-            onClick={handleClick}
-            className={`${className}`}
-            data-testid={testId}
-            {...props}
-        >
-            {children}
-        </NHSButton>
-    )
+    onClick?.(event)
+  }
+
+  return (
+    <NHSButton
+      onClick={handleClick}
+      className={`${className}`}
+      data-testid={testId}
+      {...props}
+    >
+      {children}
+    </NHSButton>
+  )
 
 }
