@@ -24,8 +24,26 @@ export const APP_CONFIG = {
 // API Endpoints
 export const API_ENDPOINTS = {
   TRACKER_USER_INFO: "/api/tracker-user-info",
-  SELECTED_ROLE: "/api/selected-role"
+  SELECTED_ROLE: "/api/selected-role",
+  PRESCRIPTION_LIST: "/api/prescription-list",
+  CIS2_SIGNOUT_ENDPOINT: "/api/cis2-signout",
+  PRESCRIPTION_DETAILS: "/api/prescription-details"
 } as const
+
+// Web page paths
+export const FRONTEND_PATHS = {
+  PRESCRIPTION_NOT_FOUND: "/prescription-not-found",
+  PRESCRIPTION_RESULTS: "/prescription-results",
+  LOGIN: "/login",
+  LOGOUT: "/logout",
+  SELECT_ROLE: "/select-role",
+  SELECTED_ROLE: "/selected-role",
+  CHANGE_ROLE: "/change-role",
+  SEARCH: "/search"
+}
+
+// This needs to be provided in backend requests as a header
+export const NHS_REQUEST_URID = "555254242106"
 
 // Type for environment
 export type MockAuthEnvironment = "dev" | "dev-pr" | "int" | "qa";
