@@ -1,12 +1,12 @@
+import {FRONTEND_PATHS} from "@/constants/environment"
+
 // HEADER strings
 export const HEADER_SERVICE = "Clinical prescription tracking service"
-export const HEADER_CONFIRM_ROLE_BUTTON = "Confirm Role"
-export const HEADER_CONFIRM_ROLE_TARGET = "/confirmrole"
 export const HEADER_CHANGE_ROLE_BUTTON = "Change Role"
-export const HEADER_CHANGE_ROLE_TARGET = "/changerole"
+export const HEADER_CHANGE_ROLE_TARGET = FRONTEND_PATHS.CHANGE_ROLE
 export const HEADER_SELECT_YOUR_ROLE_BUTTON = "Select Your Role"
-export const HEADER_SELECT_YOUR_ROLE_TARGET = "/selectyourrole"
+export const HEADER_SELECT_YOUR_ROLE_TARGET = FRONTEND_PATHS.SELECT_ROLE
 export const HEADER_PRESCRIPTION_SEARCH_BUTTON = "Search for a prescription"
-export const HEADER_PRESCRIPTION_SEARCH_TARGET = "searchforaprescription"
+export const HEADER_PRESCRIPTION_SEARCH_TARGET = FRONTEND_PATHS.SEARCH
 export const HEADER_EXIT_BUTTON = "Exit"
 export const HEADER_EXIT_TARGET = "/"
